moves tests under tests directory and cleaning up for pr
