+++ .github/workflows/test.yml
@@ -1,64 +1,70 @@
 # Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
 # See the file LICENSE for licensing terms.
 
-name: Relayer Tests
+name: Tests
 
 on:
-    push:
-        branches:
-            - "*"
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - "*"
+
+env:
+  GO_VERSION: "1.20.7"
 
 jobs:
-    build-test-relayer:
-        name: Build + Unit tests
-        runs-on: ubuntu-20.04
+  build-test-relayer:
+    name: Build + Unit tests
+    runs-on: ubuntu-20.04
 
-        steps:
-            - name: Checkout repository
-              uses: actions/checkout@v3
-              with:
-                  path: awm-relayer
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+        with:
+          path: awm-relayer
 
-            - name: Setup Go
-              uses: actions/setup-go@v4
-              with:
-                  go-version: "1.20.7"
+      - name: Setup Go
+        uses: actions/setup-go@v4
+        with:
+          go-version: $GO_VERSION
 
-            - name: Build Relayer
-              run: |
-                  cd awm-relayer
-                  go mod tidy
-                  ./scripts/build.sh
+      - name: Build Relayer
+        run: |
+          cd awm-relayer
+          go mod tidy
+          ./scripts/build.sh
 
-            - name: Run Relayer Unit Tests
-              run: |
-                  cd awm-relayer
-                  ./scripts/test.sh
+      - name: Run Relayer Unit Tests
+        run: |
+          cd awm-relayer
+          ./scripts/test.sh
 
-    e2e_tests:
-        runs-on: ubuntu-20.04
-        name: e2e_tests
+  e2e_tests:
+    runs-on: ubuntu-20.04
+    name: e2e_tests
 
-        steps:
-            - name: Checkout subnet-evm repository
-              uses: actions/checkout@v3
-              with:
-                  repository: ava-labs/subnet-evm
-                  ref: v0.5.4
+    steps:
+      - name: Checkout subnet-evm repository
+        uses: actions/checkout@v4
+        with:
+          repository: ava-labs/subnet-evm
+          ref: v0.5.4
 
-            - name: Setup Go
-              uses: actions/setup-go@v4
-              with:
-                  go-version: "1.20.7"
+      - name: Setup Go
+        uses: actions/setup-go@v4
+        with:
+          go-version: $GO_VERSION
 
-            - name: Install AvalancheGo Release
-              run: BASEDIR=/tmp/e2e-test AVALANCHEGO_BUILD_PATH=/tmp/e2e-test/avalanchego ./scripts/install_avalanchego_release.sh
+      - name: Install AvalancheGo Release
+        run: BASEDIR=/tmp/e2e-test AVALANCHEGO_BUILD_PATH=/tmp/e2e-test/avalanchego ./scripts/install_avalanchego_release.sh
 
-            - name: Build Subnet-EVM Plugin Binary
-              run: ./scripts/build.sh /tmp/e2e-test/avalanchego/plugins/srEXiWaHuhNyGwPUi444Tu47ZEDwxTWrbQiuD7FmgSAQ6X7Dy
+      - name: Build Subnet-EVM Plugin Binary
+        run: ./scripts/build.sh /tmp/e2e-test/avalanchego/plugins/srEXiWaHuhNyGwPUi444Tu47ZEDwxTWrbQiuD7FmgSAQ6X7Dy
 
-            - name: Checkout awm-relayer repository
-              uses: actions/checkout@v3
+      - name: Checkout awm-relayer repository
+        uses: actions/checkout@v3
 
-            - name: Run E2E Tests
-              run: AVALANCHEGO_BUILD_PATH=/tmp/e2e-test/avalanchego DATA_DIR=/tmp/e2e-test/data ./scripts/e2e_test.sh
+      - name: Run E2E Tests
+        run: AVALANCHEGO_BUILD_PATH=/tmp/e2e-test/avalanchego DATA_DIR=/tmp/e2e-test/data ./scripts/e2e_test.sh
