+++ scripts/e2e_test.sh
@@ -4,25 +4,26 @@
 
 set -e
 
-AWM_RELAYER_PATH=$(
+RELAYER_PATH=$(
   cd "$(dirname "${BASH_SOURCE[0]}")"
   cd .. && pwd
 )
 
-source "$AWM_RELAYER_PATH"/scripts/constants.sh
+source "$RELAYER_PATH"/scripts/constants.sh
 
-source "$AWM_RELAYER_PATH"/scripts/versions.sh
+source "$RELAYER_PATH"/scripts/versions.sh
 
 # Build ginkgo
 # to install the ginkgo binary (required for test build and run)
 go install -v github.com/onsi/ginkgo/v2/ginkgo@${GINKGO_VERSION}
 
-ACK_GINKGO_RC=true ginkgo build ./tests/e2e
+ginkgo build ./tests/
 
 # Run the tests
 echo "Running e2e tests $RUN_E2E"
-RUN_E2E=true ./tests/e2e/e2e.test \
+RUN_E2E=true ./tests/tests.test \
   --ginkgo.vv \
   --ginkgo.label-filter=${GINKGO_LABEL_FILTER:-""}
 
-  echo "e2e tests passed"
+echo "e2e tests passed"
+exit 0
