+++ tests/utils.go
@@ -1,0 +1,40 @@
+// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
+// See the file LICENSE for licensing terms.
+
+package tests
+
+import (
+	"fmt"
+	"strconv"
+	"strings"
+
+	"github.com/onsi/gomega"
+)
+
+func toWebsocketURI(uri string, blockchainID string) string {
+	return fmt.Sprintf("ws://%s/ext/bc/%s/ws", strings.TrimPrefix(uri, "http://"), blockchainID)
+}
+
+func getURIHostAndPort(uri string) (string, uint32, error) {
+	// At a minimum uri should have http:// of 7 characters
+	gomega.Expect(len(uri)).Should(gomega.BeNumerically(">", 7))
+	if uri[:7] == "http://" {
+		uri = uri[7:]
+	} else if uri[:8] == "https://" {
+		uri = uri[8:]
+	} else {
+		return "", 0, fmt.Errorf("invalid uri: %s", uri)
+	}
+
+	// Split the uri into host and port
+	hostAndPort := strings.Split(uri, ":")
+	gomega.Expect(len(hostAndPort)).Should(gomega.Equal(2))
+
+	// Parse the port
+	port, err := strconv.ParseUint(hostAndPort[1], 10, 32)
+	if err != nil {
+		return "", 0, fmt.Errorf("failed to parse port: %w", err)
+	}
+
+	return hostAndPort[0], uint32(port), nil
+}
